Some cleanup
