+++ chainstate/test-suite/src/tests/reorgs_tests.rs
@@ -49,7 +49,8 @@ fn reorg_simple(#[case] seed: Seed) {
         let genesis_id = tf.genesis().get_id();
         assert_eq!(tf.best_block_id(), genesis_id);
 
-        let block_a = tf.make_block_builder().add_test_transaction(&mut rng).build();
+        let block_a =
+            tf.make_block_builder().add_test_transaction_from_best_block(&mut rng).build();
         tf.process_block(block_a.clone(), BlockSource::Local).unwrap();
         assert_eq!(tf.best_block_id(), block_a.get_id());
